+++ server/api/routes/quotes.js
@@ -1,10 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const calculations = require('../../calculations/price_calculations')
+const calculations = require('../../quotes_service/get_quote')
 
-router.get("/", (req, res, next) => {
+router.get("/", (req, res) => {
     const {term, coverage, age, height, weight} = req.query
-    const result = calculations.calcPrice(term, coverage, age, height, weight)
+    const result = calculations.getQuote(term, coverage, age, height, weight)
     res.status(200).send(result)
 });
