+++ server/quotes_service/find_factor.js
