+++ server/quotes_service/rates_table.xlsx
